requirements.txt
# ── Image generation ──────────────────────────────────────────
# Leonardo AI — via REST API (httpx), no SDK needed

# ── Video generation ──────────────────────────────────────────
# Kling AI — via REST API (httpx), no SDK needed

# ── Caption generation ────────────────────────────────────────
openai==1.30.1

# ── HTTP (async) ──────────────────────────────────────────────
httpx==0.27.2

# ── Scheduler ─────────────────────────────────────────────────
APScheduler==3.10.4

# ── Logging ───────────────────────────────────────────────────
loguru==0.7.2

# ── Notifications ─────────────────────────────────────────────
plyer==2.1.0

# ── FastAPI dashboard ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
Pillow==10.4.0
